Move standard_error into TTFZ
This function will only be used in synthetic testing, and does not
belong in the basic TF class, since statistics about the TF
will be computed downstream with apparent resisitivities.
