+++ aurora/transfer_function/base.py
@@ -251,37 +251,3 @@ class TransferFunction(object):
         self.num_segments.data[:, index] = regression_estimator.n_data
 
         return
-
-    def standard_error(self):
-        """
-        TODO: make this a property that returns self._standard_error so it doesn't
-        compute every time you call it.
-        Returns
-        -------
-
-        """
-        stderr = np.zeros(self.tf.data.shape)
-        standard_error = xr.DataArray(
-            stderr,
-            dims=["output_channel", "input_channel", "period"],
-            coords={
-                "output_channel": self.tf_header.output_channels,
-                "input_channel": self.tf_header.input_channels,
-                "period": self.periods,
-            },
-        )
-        for out_ch in self.tf_header.output_channels:
-            for inp_ch in self.tf_header.input_channels:
-                for T in self.periods:
-                    cov_ss = self.cov_ss_inv.loc[inp_ch, inp_ch, T]
-                    cov_nn = self.cov_nn.loc[out_ch, out_ch, T]
-                    std_err = np.sqrt(np.abs(cov_ss * cov_nn))
-                    standard_error.loc[out_ch, inp_ch, T] = std_err
-
-        return standard_error
-
-    def from_emtf_zfile(self):
-        pass
-
-    def to_emtf_zfile(self):
-        pass
